fix: Final polish for v0.2.0 (Schema, Types, Paths)

+++ worker/common/src/models.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Deserialize, Serialize, Clone, PartialEq, Eq, Hash)]
 pub struct PostalCode {
     pub zip_code: String,
-    pub prefecture_id: i32,
+    pub prefecture_id: i16,
     pub city_id: String,
     pub prefecture: String,
     pub city: String,
